refactor : 로그인 기본 틀(2), JWT 의존성 버전 업
