+++ src/main/java/org/example/outsourcing_project/common/jwt/JwtProvider.java
@@ -1,0 +1,8 @@
+package org.example.outsourcing_project.common.jwt;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class JwtProvider {
+
+}
