+++ src/main/java/org/example/outsourcing_project/common/jwt/LoginUser.java
@@ -1,13 +1,10 @@
-package org.example.outsourcing_project.domain.auth.filter;
+package org.example.outsourcing_project.common.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * 로그인 필터 구현
- */
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface LoginUser {
