+++ src/main/java/org/example/outsourcing_project/domain/auth/dto/LoginRequestDto.java
@@ -1,5 +1,7 @@
 package org.example.outsourcing_project.domain.auth.dto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -7,6 +9,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public class LoginRequestDto {
 
+    @Email(message = "올바른 이메일 형식을 입력하세요.")
     private final String email;
+
+    @NotBlank(message = "비밀번호를 입력해주세요.")
     private final String password;
 }
